+++ examples/js/components/App.js
@@ -59,6 +59,9 @@ class App extends React.Component {
       text: 'Advance data edit&insert',
       href: 'advance'
     }, {
+      text: 'Expandable Row',
+      href: 'expandRow'
+    }, {
       text: 'Other',
       href: 'others'
     }, {