+++ src/store/TableDataStore.js
@@ -220,12 +220,12 @@ export class TableDataStore {
 
   addAtBegin(newObj) {
     if (!newObj[this.keyField] || newObj[this.keyField].toString() === '') {
-      throw `${this.keyField} can't be empty value.`;
+      throw new Error(`${this.keyField} can't be empty value.`);
     }
     const currentDisplayData = this.getCurrentDisplayData();
     currentDisplayData.forEach(function(row) {
       if (row[this.keyField].toString() === newObj[this.keyField].toString()) {
-        throw `${this.keyField} ${newObj[this.keyField]} already exists`;
+        throw new Error(`${this.keyField} ${newObj[this.keyField]} already exists`);
       }
     }, this);
     currentDisplayData.unshift(newObj);
@@ -237,12 +237,12 @@ export class TableDataStore {
 
   add(newObj) {
     if (!newObj[this.keyField] || newObj[this.keyField].toString() === '') {
-      throw `${this.keyField} can't be empty value.`;
+      throw new Error(`${this.keyField} can't be empty value.`);
     }
     const currentDisplayData = this.getCurrentDisplayData();
     currentDisplayData.forEach(function(row) {
       if (row[this.keyField].toString() === newObj[this.keyField].toString()) {
-        throw `${this.keyField} ${newObj[this.keyField]} already exists`;
+        throw new Error(`${this.keyField} ${newObj[this.keyField]} already exists`);
       }
     }, this);